Move ViewInspector descriptor tests to own module.
